requirements.txt
pygame==2.5.2
PyQt6==6.6.1
Pillow==10.2.0 